requirements.txt
python-dotenv
PyPDF2
faiss-cpu
tiktoken
chromadb
pypdf
pytest
pysqlite3-binary
openai==1.32.0
langchain_community==0.2.7
langchain_openai==0.1.15
langchain==0.2.7
langgraph
langchain_anthropic==0.34.2